requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
Django==2.1.15
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.0.1
django-seed==0.2.2
djangorestframework==3.11.0
Faker==4.1.0
importlib-metadata==1.6.1
numpy==1.18.5
pandas==1.0.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
zipp==3.1.0
